helix-runtime/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "helix-runtime"
version = "0.1.0"
description = "Asynchronous runtime and worker library for the Helix platform."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Alexander Shubert" },
]
dependencies = []

[project.scripts]
helix-runtime = "helix_runtime.cli:main"

[project.optional-dependencies]
brokers = [
  "kafka-python>=2.0",
  "pika>=1.3",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
